fix(client): stop externalizing @deepseek-ai/schemastery in the client bundle
The web settings card failed to load with "client-modules: require(\"@deepseek-ai/schemastery\") missed the module table": the client bundle emitted a require the loader module table cannot answer (not a platform seed word, not a shell-own module, no registered factory). The host's server half provides schemastery in-box, but the browser module table is a frozen shell-composed set — server npm deps never flow into it.

Move the Config schema into a host-only module (src/schema.ts); the client graph now reaches FallbacksConfig type-only. Drop @deepseek-ai/schemastery from CLIENT_EXTERNALS so the bundle purity gate (resolveId + emitted-surface scan) guards the split.
