+++ scripts/build-client.ts
@@ -66,11 +66,6 @@ export const CLIENT_EXTERNALS: readonly string[] = [
   '@deepseek-ai/dsh-client-ui-attachment',
   '@deepseek-ai/dsh-client-schema-form',
   '@deepseek-ai/dsh-client-runtime/client',
-  // @deepseek-ai/schemastery: config.ts's schema value import reaches the
-  // client module graph (client consumes FallbacksConfig types via config.ts);
-  // the dsh host provides the package in-box (dsh-settings depends on it), so
-  // it stays external like the other @deepseek-ai/* runtime faces.
-  '@deepseek-ai/schemastery',
 ]
 
 /**
